fix: solve pull request comments
